refactor: 스터디 프로젝트로 학습 클래스 이동
