+++ src/test/kotlin/study/caculator/TextCalculatorTest.kt
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
 import org.junit.jupiter.params.provider.NullAndEmptySource
-import study.calculator.TextCalculator
 
 internal class TextCalculatorTest {
