+++ src/test/kotlin/study/caculator/data/TextReaderTest.kt
@@ -5,7 +5,6 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.NullAndEmptySource
-import study.calculator.data.TextReader
 
 internal class TextReaderTest {
