+++ src/test/kotlin/study/caculator/mathcalculator/Calculator.kt
@@ -1,4 +1,4 @@
-package study.calculator.mathcalculator
+package study.caculator.mathcalculator
 
 interface Calculator {
     fun calculate(x: Int, y: Int): Int
