+++ src/test/kotlin/study/caculator/mathcalculator/DivideCalculator.kt
@@ -1,4 +1,4 @@
-package study.calculator.mathcalculator
+package study.caculator.mathcalculator
 
 class DivideCalculator : Calculator {
     override fun calculate(x: Int, y: Int): Int {
