+++ src/test/kotlin/study/caculator/mathcalculator/MinusCalculator.kt
@@ -1,4 +1,4 @@
-package study.calculator.mathcalculator
+package study.caculator.mathcalculator
 
 class MinusCalculator : Calculator {
     override fun calculate(x: Int, y: Int): Int {
