+++ src/test/kotlin/study/caculator/mathcalculator/MultiplyCalculator.kt
@@ -1,4 +1,4 @@
-package study.calculator.mathcalculator
+package study.caculator.mathcalculator
 
 class MultiplyCalculator : Calculator {
     override fun calculate(x: Int, y: Int): Int {
